Restructure modules to be cleaner
